requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-braces==1.13.0
django-filter==2.1.0
djangorestframework==3.9.1
gunicorn==19.9.0
Markdown==3.0.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
whitenoise==4.1.2
